Update hue_on
added more comments about how and when this is used and how. Switched the order of GUI and CLI notifications to ensure the CLI notification one is noticed, especially if the command is run in a terminal on a machine without notify-send installed, so that error message does not get the attention.