System change, config now for manjaro i3
After switching to SSD in my X200 i went back to Manjaro I3 instead of debian. That means new configs, for pretty much the same setup. Differences are in keybindings and programs. I'll try to upload as many scripts of those that I use, and link to the ones i've been lucky to find on my way. :)